# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Patchman-engine API",
        "description": "API of the Patch application on [console.redhat.com](https://console.redhat.com)\n\nSyntax of the `filter[name]` query parameters is described in  [Filters documentation](https://github.com/RedHatInsights/patchman-engine/wiki/API-custom-filters)",
        "contact": {},
        "license": {
            "name": "GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "v1.18.52"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/patch/v1/baselines/{baseline_id}": {
            "get": {
                "summary": "Show baseline detail by given baseline ID",
                "description": "Show baseline detail by given baseline ID",
                "operationId": "detailBaseline",
                "parameters": [
                    {
                        "name": "baseline_id",
                        "in": "path",
                        "description": "Baseline ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.BaselineDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            },
            "put": {
                "summary": "Update a baseline for my set of systems",
                "description": "Update a baseline for my set of systems",
                "operationId": "updateBaseline",
                "parameters": [
                    {
                        "name": "baseline_id",
                        "in": "path",
                        "description": "Baseline ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/controllers.UpdateBaselineRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.UpdateBaselineResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ],
                "x-codegen-request-body-name": "body"
            },
            "delete": {
                "summary": "Delete a baseline",
                "description": "Delete a baseline",
                "operationId": "baselineDelete",
                "parameters": [
                    {
                        "name": "baseline_id",
                        "in": "path",
                        "description": "Baseline ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.DeleteBaselineResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            }
        },
        "/api/patch/v1/baselines/{baseline_id}/systems": {
            "get": {
                "summary": "Show me all systems belonging to a baseline",
                "description": "Show me all systems applicable to a baseline",
                "operationId": "listBaselineSystems",
                "parameters": [
                    {
                        "name": "baseline_id",
                        "in": "path",
                        "description": "Baseline ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit for paging, set -1 to return all",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for paging",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "config"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Find matching text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[display_name]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tag filter",
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.BaselineSystemsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "controllers.BaselineConfig": {
                "type": "object",
                "properties": {
                    "to_time": {
                        "type": "string",
                        "description": "Filter applicable advisories (updates) by the latest publish time.",
                        "example": "2022-12-31T12:00:00-04:00"
                    }
                }
            },
            "controllers.BaselineDetailAttributes": {
                "type": "object",
                "properties": {
                    "config": {
                        "$ref": "#/components/schemas/controllers.BaselineConfig"
                    },
                    "name": {
                        "type": "string",
                        "description": "Baseline name",
                        "example": "my_baseline"
                    }
                }
            },
            "controllers.BaselineDetailItem": {
                "type": "object",
                "properties": {
                    "attributes": {
                        "$ref": "#/components/schemas/controllers.BaselineDetailAttributes"
                    },
                    "id": {
                        "type": "integer",
                        "description": "Baseline ID",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "description": "Document type name",
                        "example": "baseline"
                    }
                }
            },
            "controllers.BaselineDetailResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/controllers.BaselineDetailItem"
                    }
                }
            },
...
            "controllers.BaselineSystemAttributes": {
                "type": "object",
                "properties": {
                    "display_name": {
                        "type": "string",
                        "description": "Baseline system display name",
                        "example": "my-baselined-system"
                    }
                }
            },
            "controllers.BaselineSystemItem": {
                "type": "object",
                "properties": {
                    "attributes": {
                        "$ref": "#/components/schemas/controllers.BaselineSystemAttributes"
                    },
                    "inventory_id": {
                        "type": "string",
                        "description": "Baseline system inventory ID (uuid format)",
                        "example": "00000000-0000-0000-0000-000000000001"
                    },
                    "type": {
                        "type": "string",
                        "description": "Document type name",
                        "example": "baseline_system"
                    }
                }
            },
            "controllers.BaselineSystemsResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/controllers.BaselineSystemItem"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/controllers.Links"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/controllers.ListMeta"
                    }
                }
            },
...
            "controllers.DeleteBaselineResponse": {
                "type": "object",
                "properties": {
                    "baseline_id": {
                        "type": "integer",
                        "description": "Updated baseline unique ID, it can not be changed",
                        "example": 1
                    }
                }
            },
            "controllers.FilterData": {
                "type": "object",
                "properties": {
                    "op": {
                        "type": "string"
                    },
                    "values": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "controllers.Links": {
                "type": "object",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "/link/to/the/first"
                    },
                    "last": {
                        "type": "string",
                        "example": "/link/to/the/last"
                    },
                    "next": {
                        "type": "string",
                        "example": "/link/to/the/next"
                    },
                    "previous": {
                        "type": "string",
                        "example": "/link/to/the/previous"
                    }
                }
            },
            "controllers.ListMeta": {
                "type": "object",
                "properties": {
                    "filter": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/controllers.FilterData"
                        },
                        "description": "Used filters"
                    },
                    "limit": {
                        "type": "integer",
                        "description": "Used response limit (page size) - pagination",
                        "example": 20
                    },
                    "offset": {
                        "type": "integer",
                        "description": "Used response offset - pagination",
                        "example": 0
                    },
                    "search": {
                        "type": "string",
                        "description": "Used search terms",
                        "example": "kernel"
                    },
                    "sort": {
                        "type": "array",
                        "description": "Used sorting fields",
                        "example": [
                            "name"
                        ],
                        "items": {
                            "type": "string"
                        }
                    },
                    "subtotals": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "description": "Some subtotals used by some endpoints"
                    },
                    "total_items": {
                        "type": "integer",
                        "description": "Total items count to return",
                        "example": 1000
                    }
                }
            },
...
            "controllers.UpdateBaselineRequest": {
                "type": "object",
                "properties": {
                    "config": {
                        "$ref": "#/components/schemas/controllers.BaselineConfig"
                    },
                    "inventory_ids": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "boolean"
                        },
                        "description": "Map of inventories to add to (true) or remove (false) from given baseline (optional)"
                    },
                    "name": {
                        "type": "string",
                        "description": "Updated baseline name (optional)",
                        "example": "my-changed-baseline-name"
                    }
                }
            },
            "controllers.UpdateBaselineResponse": {
                "type": "object",
                "properties": {
                    "baseline_id": {
                        "type": "integer",
                        "description": "Updated baseline unique ID, it can not be changed",
                        "example": 1
                    }
                }
            },
...
            "utils.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "RhIdentity": {
                "type": "apiKey",
                "name": "x-rh-identity",
                "in": "header"
            }
        }
    }
}